+++ src/main/resources/jscoverage-localstorage.js
@@ -1,4 +1,4 @@
-if (typeof(_$jscoverage) === "undefined" && (typeof(Storage) !== "undefined") && localStorage["jscover"])
+if (typeof(_$jscoverage) === "undefined" && (typeof(Storage) !== "undefined") && typeof(localStorage["jscover"]) !== "undefined")
     _$jscoverage = jscoverage_parseCoverageJSON(localStorage["jscover"]);
 if (typeof(jscoverbeforeunload) === "undefined") {
     var jscoverbeforeunload = (window.onbeforeunload) ? window.onbeforeunload : function () {};