+++ src/test-acceptance/java/jscover/server/HtmlUnitLocalStorageTest.java
@@ -340,27 +340,25 @@ library.  If this is what you want to do, use the GNU Lesser General
 Public License instead of this License.
  */
 
-package jscover;
+package jscover.server;
 
-import org.junit.After;
+import com.gargoylesoftware.htmlunit.BrowserVersion;
+import com.gargoylesoftware.htmlunit.WebClient;
+import com.gargoylesoftware.htmlunit.html.HtmlElement;
+import com.gargoylesoftware.htmlunit.html.HtmlPage;
+import jscover.Main;
 import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.phantomjs.PhantomJSDriver;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
 
 import java.io.File;
 import java.io.IOException;
 
 import static org.junit.Assert.assertEquals;
-import static org.openqa.selenium.support.ui.ExpectedConditions.textToBePresentInElementLocated;
 
-public class WebDriverLocalStorageTest {
+public class HtmlUnitLocalStorageTest {
     private static Thread server;
 
-    protected WebDriver webClient = getWebClient();
+    protected WebClient webClient = new WebClient(BrowserVersion.FIREFOX_24);
     private String[] args = new String[]{
             "-ws",
             "--document-root=src/test-acceptance/resources",
@@ -369,10 +367,6 @@ public class WebDriverLocalStorageTest {
             "--local-storage",
             "--report-dir=" + getReportDir()
     };
-
-    protected WebDriver getWebClient() {
-        return new PhantomJSDriver();
-    }
 
     protected String getReportDir() {
         return "target/localStorage";
@@ -397,20 +391,6 @@ public class WebDriverLocalStorageTest {
         }
     }
 
-    @After
-    public void tearDown() {
-        try {
-            webClient.close();
-        } catch (Throwable t) {
-            t.printStackTrace();
-        }
-        try {
-            webClient.quit();
-        } catch (Throwable t) {
-            t.printStackTrace();
-        }
-    }
-
     protected String[] getArgs() {
         return args;
     }
@@ -421,75 +401,59 @@ public class WebDriverLocalStorageTest {
 
     @Test
     public void shouldStoreCoverageDataInLocalStorage() throws IOException, InterruptedException {
-        webClient.get("http://localhost:9001/jscoverage-clear-local-storage.html");
-        new WebDriverWait(webClient, 1).until(ExpectedConditions.textToBePresentInElementLocated(By.id("deleteLocalStorage"), "Deleted localStorage['jscover']"));
-        webClient.get("http://localhost:9001/" + getTestUrl());
-        webClient.get("http://localhost:9001/jscoverage.html");
-        verifyTotal(webClient, 15, 0, 0);
+        webClient.getPage("http://localhost:9001/jscoverage-clear-local-storage.html");
+        webClient.getPage("http://localhost:9001/" + getTestUrl());
+        HtmlPage page = webClient.getPage("http://localhost:9001/jscoverage.html");
+        verifyTotal(page, 15, 0, 0);
 
         clickItem(1, 57, 12, 33);
         clickItem(2, 73, 37, 66);
         clickItem(3, 84, 62, 66);
         clickItem(4, 100, 87, 100);
 
-        new WebDriverWait(webClient, 1).until(ExpectedConditions.elementToBeClickable(By.id("storeTab")));
-        webClient.findElement(By.id("storeTab")).click();
-
-        new WebDriverWait(webClient, 1).until(ExpectedConditions.textToBePresentInElementLocated(By.id("progressLabel"), "Done"));
-        new WebDriverWait(webClient, 1).until(ExpectedConditions.elementToBeClickable(By.id("storeButton")));
-        Thread.sleep(100);
-        webClient.findElement(By.id("storeButton")).click();
-        new WebDriverWait(webClient, 2).until(ExpectedConditions.textToBePresentInElementLocated(By.id("storeDiv"), "Coverage data stored at"));
+        //page.getHtmlElementById("storeTab").click();
+        //((HtmlElement)page.getElementById("storeButton")).click();
     }
 
     @Test
     public void shouldClearLocalStorage() throws IOException, InterruptedException {
-        webClient.get("http://localhost:9001/jscoverage-clear-local-storage.html");
-        new WebDriverWait(webClient, 1).until(ExpectedConditions.textToBePresentInElementLocated(By.id("deleteLocalStorage"), "Deleted localStorage['jscover']"));
-        webClient.get("http://localhost:9001/" + getTestUrl());
-        webClient.get("http://localhost:9001/jscoverage.html");
-        verifyTotal(webClient, 15, 0, 0);
+        webClient.getPage("http://localhost:9001/jscoverage-clear-local-storage.html");
+        webClient.getPage("http://localhost:9001/" + getTestUrl());
+        HtmlPage page = webClient.getPage("http://localhost:9001/jscoverage.html");
+        verifyTotal(page, 15, 0, 0);
 
         clickItem(1, 57, 12, 33);
         clickItem(2, 73, 37, 66);
-        webClient.get("http://localhost:9001/jscoverage-clear-local-storage.html");
-        new WebDriverWait(webClient, 1).until(ExpectedConditions.textToBePresentInElementLocated(By.id("deleteLocalStorage"), "Deleted localStorage['jscover']"));
+        webClient.getPage("http://localhost:9001/jscoverage-clear-local-storage.html");
 
-        webClient.get("http://localhost:9001/" + getTestUrl());
-        webClient.get("http://localhost:9001/jscoverage.html");
-        verifyTotal(webClient, 15, 0, 0);
+        webClient.getPage("http://localhost:9001/" + getTestUrl());
+        page = webClient.getPage("http://localhost:9001/jscoverage.html");
+        verifyTotal(page, 15, 0, 0);
         clickItem(1, 57, 12, 33);
         clickItem(2, 73, 37, 66);
         clickItem(3, 84, 62, 66);
         clickItem(4, 100, 87, 100);
 
-        new WebDriverWait(webClient, 1).until(ExpectedConditions.elementToBeClickable(By.id("storeTab")));
-        webClient.findElement(By.id("storeTab")).click();
-
-        new WebDriverWait(webClient, 1).until(ExpectedConditions.textToBePresentInElementLocated(By.id("progressLabel"), "Done"));
-        new WebDriverWait(webClient, 1).until(ExpectedConditions.elementToBeClickable(By.id("storeButton")));
-        Thread.sleep(100);
-        webClient.findElement(By.id("storeButton")).click();
-        new WebDriverWait(webClient, 2).until(ExpectedConditions.textToBePresentInElementLocated(By.id("storeDiv"), "Coverage data stored at"));
+        //((HtmlElement)page.getElementById("storeTab")).click();
+        //((HtmlElement)page.getElementById("storeButton")).click();
     }
 
     private void clickItem(int id, int linePercentage, int branchPercentage, int functionPercentage) throws IOException {
-        webClient.get("http://localhost:9001/" + getTestUrl());
-        webClient.findElement(By.id("radio" + id)).click();
-        webClient.get("http://localhost:9001/jscoverage.html");
-        verifyTotal(webClient, linePercentage, branchPercentage, functionPercentage);
+        HtmlPage page = webClient.getPage("http://localhost:9001/" + getTestUrl());
+        ((HtmlElement)page.getElementById("radio" + id)).click();
+        page = webClient.getPage("http://localhost:9001/jscoverage.html");
+        verifyTotal(page, linePercentage, branchPercentage, functionPercentage);
     }
 
-    protected void verifyTotal(WebDriver webClient, int percentage, int branchPercentage, int functionPercentage) throws IOException {
-        webClient.findElement(By.id("summaryTab")).click();
-
-        verifyTotals(webClient, percentage, branchPercentage, functionPercentage);
+    protected void verifyTotal(HtmlPage page, int percentage, int branchPercentage, int functionPercentage) throws IOException {
+        page = ((HtmlElement)page.getElementById("summaryTab")).click();
+        webClient.waitForBackgroundJavaScript(2000);
+        verifyTotals(page, percentage, branchPercentage, functionPercentage);
     }
 
-    protected void verifyTotals(WebDriver webClient, int percentage, int branchPercentage, int functionPercentage) {
-        new WebDriverWait(webClient, 1).until(textToBePresentInElementLocated(By.id("summaryTotal"), "%"));
-        assertEquals(percentage + "%", webClient.findElement(By.id("summaryTotal")).getText());
-        assertEquals(branchPercentage + "%", webClient.findElement(By.id("branchSummaryTotal")).getText());
-        assertEquals(functionPercentage + "%", webClient.findElement(By.id("functionSummaryTotal")).getText());
+    protected void verifyTotals(HtmlPage page, int percentage, int branchPercentage, int functionPercentage) {
+        assertEquals(percentage + "%", page.getElementById("summaryTotal").getTextContent());
+        assertEquals(branchPercentage + "%", page.getElementById("branchSummaryTotal").getTextContent());
+        assertEquals(functionPercentage + "%", page.getElementById("functionSummaryTotal").getTextContent());
     }
 }